requirements.txt
# NiChart packages (not generally needed anymore thanks to Docker based deployment, but kept here for reference)
#spare-scores
#NiChartHarmonize
#DLICV
#DLMUSE
#NiChart_DLMUSE>=1.0.9
#DLWMLS @ git+https://github.com/CBICA/DLWMLS
NiChart_common_utils

# general python dependencies
torch==2.2.1
argparse
nibabel
pulp==2.7.0
tabulate==0.8.10
sphinx
pydantic
pyyaml
humanize
graphviz
dicom2nifti

# streamlit
streamlit==1.50.0
st_pages==1.0.1
streamlit_plotly_events
streamlit-base-extras
streamlit-extras
streamlit-autorefresh
streamlit-option-menu
streamlit-image-select
streamlit-antd-components 
pycountry
chart_studio
Jinja2
plotly-express
stqdm
pyjwt
docker

# chatbot 
boto3
